perf(internal): streamline compat module
We no longer need to define many compatibility aliases, so we can
streamline the content of the compat module for a faster load and
decreased startup time.
