+++ ddtrace/_trace/span.py
@@ -1,3 +1,4 @@
+from io import StringIO
 import math
 import pprint
 import sys
@@ -44,7 +45,6 @@ from ddtrace.internal import core
 from ddtrace.internal._rand import rand64bits as _rand64bits
 from ddtrace.internal._rand import rand128bits as _rand128bits
 from ddtrace.internal.compat import NumericType
-from ddtrace.internal.compat import StringIO
 from ddtrace.internal.compat import ensure_text
 from ddtrace.internal.compat import is_integer
 from ddtrace.internal.constants import MAX_UINT_64BITS as _MAX_UINT_64BITS
