+++ ddtrace/contrib/internal/django/patch.py
@@ -7,6 +7,7 @@ Django internals are instrumented via normal `patch()`.
 specific Django apps like Django Rest Framework (DRF).
 """
 
+from collections.abc import Iterable
 import functools
 from inspect import getmro
 from inspect import isclass
@@ -32,7 +33,6 @@ from ddtrace.ext import net
 from ddtrace.ext import sql as sqlx
 from ddtrace.internal import core
 from ddtrace.internal._exceptions import BlockingException
-from ddtrace.internal.compat import Iterable
 from ddtrace.internal.compat import maybe_stringify
 from ddtrace.internal.constants import COMPONENT
 from ddtrace.internal.core.event_hub import ResultType
