+++ ddtrace/contrib/internal/httplib/patch.py
@@ -1,6 +1,8 @@
 import functools
+import http.client as httplib
 import os
 import sys
+from urllib import parse
 
 import wrapt
 
@@ -11,8 +13,6 @@ from ddtrace.contrib import trace_utils
 from ddtrace.contrib.internal.trace_utils import unwrap as _u
 from ddtrace.ext import SpanKind
 from ddtrace.ext import SpanTypes
-from ddtrace.internal.compat import httplib
-from ddtrace.internal.compat import parse
 from ddtrace.internal.constants import _HTTPLIB_NO_TRACE_REQUEST
 from ddtrace.internal.constants import COMPONENT
 from ddtrace.internal.logger import get_logger
