+++ ddtrace/contrib/internal/pylibmc/client.py
@@ -1,3 +1,4 @@
+from collections.abc import Iterable
 from contextlib import contextmanager
 import random
 
@@ -16,7 +17,6 @@ from ddtrace.ext import SpanTypes
 from ddtrace.ext import db
 from ddtrace.ext import memcached
 from ddtrace.ext import net
-from ddtrace.internal.compat import Iterable
 from ddtrace.internal.constants import COMPONENT
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.schema import schematize_cache_operation
