+++ ddtrace/debugging/_uploader.py
@@ -9,9 +9,9 @@ from ddtrace.debugging._encoding import LogSignalJsonEncoder
 from ddtrace.debugging._encoding import SignalQueue
 from ddtrace.debugging._metrics import metrics
 from ddtrace.debugging._signal.collector import SignalCollector
-from ddtrace.internal import compat
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.periodic import ForksafeAwakeablePeriodicService
+from ddtrace.internal.utils import http
 from ddtrace.internal.utils.http import connector
 from ddtrace.internal.utils.retry import fibonacci_backoff_with_jitter
 
@@ -81,7 +81,7 @@ class LogsIntakeUploaderV1(ForksafeAwakeablePeriodicService):
                     payload,
                     headers=self._headers,
                 )
-                resp = compat.get_connection_response(conn)
+                resp = http.get_connection_response(conn)
                 if not (200 <= resp.status < 300):
                     log.error("Failed to upload payload: [%d] %r", resp.status, resp.read())
                     meter.increment("upload.error", tags={"status": str(resp.status)})
