+++ ddtrace/internal/ci_visibility/git_client.py
@@ -25,13 +25,12 @@ from ddtrace.ext.git import extract_git_version
 from ddtrace.ext.git import extract_remote_url
 from ddtrace.ext.git import extract_workspace_path
 from ddtrace.internal.agent import get_trace_url
-from ddtrace.internal.compat import JSONDecodeError
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.utils.retry import fibonacci_backoff_with_jitter
 from ddtrace.trace import Tracer  # noqa: F401
 
-from .. import compat
 from .. import telemetry
+from ..utils import http
 from ..utils.http import Response
 from ..utils.http import get_connection
 from ..utils.http import verify_url
@@ -291,7 +290,7 @@ class CIVisibilityGitClient(object):
                 try:
                     result = serializer.search_commits_decode(response.body)
                     return result
-                except JSONDecodeError:
+                except json.JSONDecodeError:
                     request_error = ERROR_TYPES.BAD_JSON
                     log.warning("Error searching commits, response not parsable: %s", response.body)
                     return None
@@ -318,7 +317,7 @@ class CIVisibilityGitClient(object):
             conn = get_connection(url, timeout=timeout)
             log.debug("Sending request: %s %s %s %s", "POST", url_path, payload, _headers)
             conn.request("POST", url_path, payload, _headers)
-            resp = compat.get_connection_response(conn)
+            resp = http.get_connection_response(conn)
             log.debug("Response status: %s", resp.status)
             result = Response.from_http_response(resp)
         finally:
@@ -468,7 +467,7 @@ class CIVisibilityGitClientSerializerV1(object):
             return [item["id"] for item in parsed["data"] if item["type"] == "commit"]
         except KeyError:
             log.warning("Expected information not found in search_commits response", exc_info=True)
-        except JSONDecodeError:
+        except json.JSONDecodeError:
             log.warning("Unexpected decode error in search_commits response", exc_info=True)
 
         return res
