+++ ddtrace/internal/compat.py
@@ -1,21 +1,8 @@
-from collections import Counter
-from collections import OrderedDict
-from collections import defaultdict
-from collections import deque
 import functools
-from inspect import iscoroutinefunction
-from inspect import isgeneratorfunction
 import ipaddress
-import os
-import platform
 import re
 import sys
-from tempfile import mkdtemp
 import threading
-from types import BuiltinFunctionType
-from types import BuiltinMethodType
-from types import FunctionType
-from types import MethodType
 from types import TracebackType
 from typing import Any  # noqa:F401
 from typing import AnyStr  # noqa:F401
@@ -24,33 +11,13 @@ from typing import Text  # noqa:F401
 from typing import Tuple  # noqa:F401
 from typing import Type  # noqa:F401
 from typing import Union  # noqa:F401
-import warnings
-
-from wrapt.wrappers import BoundFunctionWrapper
-from wrapt.wrappers import FunctionWrapper
 
 
 __all__ = [
-    "httplib",
-    "iscoroutinefunction",
-    "Queue",
-    "StringIO",
-    "urlencode",
-    "parse",
     "maybe_stringify",
 ]
 
 PYTHON_VERSION_INFO = sys.version_info
-
-# Infos about python passed to the trace agent through the header
-PYTHON_VERSION = platform.python_version()
-PYTHON_INTERPRETER = platform.python_implementation()
-
-import http.client as httplib  # noqa: E402
-from io import StringIO  # noqa: E402
-from queue import Queue  # noqa: E402
-import urllib.parse as parse  # noqa: E402
-from urllib.parse import urlencode  # noqa: E402
 
 
 def ensure_text(s, encoding="utf-8", errors="ignore") -> str:
@@ -73,26 +40,6 @@ NumericType = Union[int, float]
 
 # Pattern class generated by `re.compile`
 pattern_type = re.Pattern
-
-try:
-    from inspect import getfullargspec  # noqa:F401
-
-    def is_not_void_function(f, argspec):
-        return (
-            argspec.args
-            or argspec.varargs
-            or argspec.varkw
-            or argspec.defaults
-            or argspec.kwonlyargs
-            or argspec.kwonlydefaults
-            or isgeneratorfunction(f)
-        )
-
-except ImportError:
-    from inspect import getargspec as getfullargspec  # type: ignore[assignment]  # noqa: F401
-
-    def is_not_void_function(f, argspec):
-        return argspec.args or argspec.varargs or argspec.keywords or argspec.defaults or isgeneratorfunction(f)
 
 
 def is_integer(obj):
@@ -151,30 +98,6 @@ def to_unicode(s):
     return str(s)
 
 
-def get_connection_response(
-    conn,  # type: httplib.HTTPConnection
-):
-    # type: (...) -> httplib.HTTPResponse
-    """Returns the response for a connection.
-
-    If using Python 2 enable buffering.
-
-    Python 2 does not enable buffering by default resulting in many recv
-    syscalls.
-
-    See:
-    https://bugs.python.org/issue4879
-    https://github.com/python/cpython/commit/3c43fcba8b67ea0cec4a443c755ce5f25990a6cf
-    """
-    return conn.getresponse()
-
-
-try:
-    from collections.abc import Iterable  # noqa:F401
-except ImportError:
-    from collections import Iterable  # type: ignore[no-redef, attr-defined]  # noqa:F401
-
-
 def maybe_stringify(obj):
     # type: (Any) -> Optional[str]
     if obj is not None:
@@ -182,52 +105,7 @@ def maybe_stringify(obj):
     return None
 
 
-NoneType = type(None)
-
-BUILTIN_SIMPLE_TYPES = frozenset([int, float, str, bytes, bool, NoneType, type, complex])
-BUILTIN_MAPPING_TYPES = frozenset([dict, defaultdict, Counter, OrderedDict])
-BUILTIN_SEQUENCE_TYPES = frozenset([list, tuple, set, frozenset, deque])
-BUILTIN_CONTAINER_TYPES = BUILTIN_MAPPING_TYPES | BUILTIN_SEQUENCE_TYPES
-BUILTIN_TYPES = BUILTIN_SIMPLE_TYPES | BUILTIN_CONTAINER_TYPES
-
-
-try:
-    from types import MethodWrapperType
-
-except ImportError:
-    MethodWrapperType = object().__init__.__class__  # type: ignore[misc]
-
-CALLABLE_TYPES = (
-    BuiltinMethodType,
-    BuiltinFunctionType,
-    FunctionType,
-    MethodType,
-    MethodWrapperType,
-    FunctionWrapper,
-    BoundFunctionWrapper,
-    property,
-    classmethod,
-    staticmethod,
-)
-BUILTIN = "builtins"
-
-
-try:
-    from typing import Collection  # noqa:F401
-except ImportError:
-    from typing import List  # noqa:F401
-    from typing import Set  # noqa:F401
-    from typing import Union  # noqa:F401
-
-    Collection = Union[List, Set, Tuple]  # type: ignore[misc,assignment]
-
 ExcInfoType = Union[Tuple[Type[BaseException], BaseException, Optional[TracebackType]], Tuple[None, None, None]]
-
-
-try:
-    from json import JSONDecodeError
-except ImportError:
-    JSONDecodeError = ValueError  # type: ignore[misc,assignment]
 
 
 def is_valid_ip(ip: str) -> bool:
@@ -249,129 +127,9 @@ def ip_is_global(ip):
     return parsed_ip.is_global
 
 
-# https://stackoverflow.com/a/19299884
-class TemporaryDirectory(object):
-    """Create and return a temporary directory.  This has the same
-    behavior as mkdtemp but can be used as a context manager.  For
-    example:
-
-        with TemporaryDirectory() as tmpdir:
-            ...
-
-    Upon exiting the context, the directory and everything contained
-    in it are removed.
-    """
-
-    def __init__(self, suffix="", prefix="tmp", _dir=None):
-        self._closed = False
-        self.name = None  # Handle mkdtemp raising an exception
-        self.name = mkdtemp(suffix, prefix, _dir)
-
-    def __repr__(self):
-        return "<{} {!r}>".format(self.__class__.__name__, self.name)
-
-    def __enter__(self):
-        return self.name
-
-    def cleanup(self, _warn=False):
-        if self.name and not self._closed:
-            try:
-                self._rmtree(self.name)
-            except (TypeError, AttributeError) as ex:
-                # Issue #10188: Emit a warning on stderr
-                # if the directory could not be cleaned
-                # up due to missing globals
-                if "None" not in str(ex):
-                    raise
-                return
-            self._closed = True
-            if _warn:
-                self._warn("Implicitly cleaning up {!r}".format(self), ResourceWarning)
-
-    def __exit__(self, exc, value, tb):
-        self.cleanup()
-
-    def __del__(self):
-        # Issue a ResourceWarning if implicit cleanup needed
-        self.cleanup(_warn=True)
-
-    # XXX (ncoghlan): The following code attempts to make
-    # this class tolerant of the module nulling out process
-    # that happens during CPython interpreter shutdown
-    # Alas, it doesn't actually manage it. See issue #10188
-    _listdir = staticmethod(os.listdir)
-    _path_join = staticmethod(os.path.join)
-    _isdir = staticmethod(os.path.isdir)
-    _islink = staticmethod(os.path.islink)
-    _remove = staticmethod(os.remove)
-    _rmdir = staticmethod(os.rmdir)
-    _warn = warnings.warn
-
-    def _rmtree(self, path):
-        # Essentially a stripped down version of shutil.rmtree.  We can't
-        # use globals because they may be None'ed out at shutdown.
-        for name in self._listdir(path):
-            fullname = self._path_join(path, name)
-            try:
-                isdir = self._isdir(fullname) and not self._islink(fullname)
-            except OSError:
-                isdir = False
-            if isdir:
-                self._rmtree(fullname)
-            else:
-                try:
-                    self._remove(fullname)
-                except OSError:
-                    pass
-        try:
-            self._rmdir(path)
-        except OSError:
-            pass
-
-
-try:
-    from shlex import quote as shquote
-except ImportError:
-    import re
-
-    _find_unsafe = re.compile(r"[^\w@%+=:,./-]").search
-
-    def shquote(s):
-        # type: (str) -> str
-        """Return a shell-escaped version of the string *s*."""
-        if not s:
-            return "''"
-        if _find_unsafe(s) is None:
-            return s
-
-        # use single quotes, and put single quotes into double quotes
-        # the string $'b is then quoted as '$'"'"'b'
-        return "'" + s.replace("'", "'\"'\"'") + "'"
-
-
-try:
-    from shlex import join as shjoin
-except ImportError:
-
-    def shjoin(args):  # type: ignore[misc]
-        # type: (Iterable[str]) -> str
-        """Return a shell-escaped string from *args*."""
-        return " ".join(shquote(arg) for arg in args)
-
-
-try:
-    from contextlib import nullcontext
-except ImportError:
-    from contextlib import contextmanager
-
-    @contextmanager  # type: ignore[no-redef]
-    def nullcontext(enter_result=None):
-        yield enter_result
-
-
-if PYTHON_VERSION_INFO >= (3, 9):
+if PYTHON_VERSION_INFO >= (3, 9, 8):
     from functools import singledispatchmethod
-elif PYTHON_VERSION_INFO >= (3, 8):
+else:
     # This fix was not backported to 3.8
     # https://github.com/python/cpython/issues/83860
     from functools import singledispatchmethod
@@ -382,42 +140,6 @@ elif PYTHON_VERSION_INFO >= (3, 8):
         return self.dispatcher.register(cls, func=method)
 
     singledispatchmethod.register = _register  # type: ignore[assignment]
-else:
-    from functools import singledispatch
-    from functools import update_wrapper
-
-    class singledispatchmethod:  # type: ignore[no-redef]
-        """Single-dispatch generic method descriptor.
-
-        Supports wrapping existing descriptors and handles non-descriptor
-        callables as instance methods.
-        """
-
-        def __init__(self, func):
-            if not callable(func) and not hasattr(func, "__get__"):
-                raise TypeError(f"{func!r} is not callable or a descriptor")
-
-            self.dispatcher = singledispatch(func)
-            self.func = func
-
-        def register(self, cls, method=None):
-            if hasattr(cls, "__func__"):
-                setattr(cls, "__annotations__", cls.__func__.__annotations__)
-            return self.dispatcher.register(cls, func=method)
-
-        def __get__(self, obj, cls=None):
-            def _method(*args, **kwargs):
-                method = self.dispatcher.dispatch(args[0].__class__)
-                return method.__get__(obj, cls)(*args, **kwargs)
-
-            _method.__isabstractmethod__ = self.__isabstractmethod__
-            _method.register = self.register
-            update_wrapper(_method, self.func)
-            return _method
-
-        @property
-        def __isabstractmethod__(self):
-            return getattr(self.func, "__isabstractmethod__", False)
 
 
 if PYTHON_VERSION_INFO >= (3, 9):
@@ -441,3 +163,21 @@ def get_mp_context():
     import multiprocessing
 
     return multiprocessing.get_context("fork" if sys.platform != "win32" else "spawn")
+
+
+def __getattr__(name: str) -> Any:
+    # These attributes are expensive to pre-compute, so we make them lazy
+    if name == "PYTHON_VERSION":
+        from platform import python_version
+
+        globals()[name] = python_version()
+
+    elif name == "PYTHON_INTERPRETER":
+        from platform import python_implementation
+
+        globals()[name] = python_implementation()
+
+    try:
+        return globals()[name]
+    except KeyError:
+        raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
