+++ ddtrace/internal/datastreams/processor.py
@@ -25,11 +25,11 @@ from ddtrace.internal.utils.retry import fibonacci_backoff_with_jitter
 
 from .._encoding import packb
 from ..agent import get_connection
-from ..compat import get_connection_response
 from ..forksafe import Lock
 from ..hostname import get_hostname
 from ..logger import get_logger
 from ..periodic import PeriodicService
+from ..utils.http import get_connection_response
 from ..writer import _human_size
 from .encoding import decode_var_int_64
 from .encoding import encode_var_int_64
