+++ ddtrace/internal/symbol_db/symbols.py
@@ -24,7 +24,6 @@ from types import ModuleType
 import typing as t
 
 from ddtrace import config
-from ddtrace.internal import compat
 from ddtrace.internal import packages
 from ddtrace.internal.agent import get_trace_url
 from ddtrace.internal.compat import singledispatchmethod
@@ -34,6 +33,7 @@ from ddtrace.internal.module import BaseModuleWatchdog
 from ddtrace.internal.module import origin
 from ddtrace.internal.runtime import get_runtime_id
 from ddtrace.internal.safety import _isinstance
+from ddtrace.internal.utils import http
 from ddtrace.internal.utils.cache import cached
 from ddtrace.internal.utils.http import FormData
 from ddtrace.internal.utils.http import connector
@@ -501,7 +501,7 @@ class ScopeContext:
             log.debug("[PID %d] SymDB: Uploading symbols payload", os.getpid())
             conn.request("POST", "/symdb/v1/input", body, headers)
 
-            return compat.get_connection_response(conn)
+            return http.get_connection_response(conn)
 
     def __bool__(self) -> bool:
         return bool(self._scopes)
