+++ tests/appsec/integrations/flask_tests/test_appsec_flask_telemetry.py
@@ -1,8 +1,9 @@
+from urllib.parse import urlencode
+
 import pytest
 
 from ddtrace.appsec._utils import get_triggers
 from ddtrace.internal import core
-from ddtrace.internal.compat import urlencode
 from ddtrace.internal.constants import BLOCKED_RESPONSE_JSON
 from tests.appsec.appsec.test_telemetry import _assert_generate_metrics
 import tests.appsec.rules as rules
