+++ tests/appsec/integrations/flask_tests/test_flask_remoteconfig.py
@@ -1,17 +1,17 @@
 import base64
 import datetime
 import hashlib
+import http.client as httplib
 import json
 import os
 import signal
 import sys
 import time
+from urllib import parse
 import uuid
 
 import pytest
 
-from ddtrace.internal.compat import httplib
-from ddtrace.internal.compat import parse
 from ddtrace.trace import tracer
 from tests.appsec.appsec_utils import gunicorn_server
 from tests.appsec.integrations.flask_tests.utils import _PORT
