+++ tests/contrib/djangorestframework/test_appsec.py
@@ -1,9 +1,10 @@
+from urllib.parse import urlencode
+
 import django
 import pytest
 
 from ddtrace.appsec._utils import get_triggers
 from ddtrace.internal import core
-from ddtrace.internal.compat import urlencode
 from tests.utils import assert_span_http_status_code
 from tests.utils import override_global_config
