+++ tests/contrib/flask/test_flask_helpers.py
@@ -1,10 +1,10 @@
 from io import BytesIO
+from io import StringIO
 
 import flask
 
 from ddtrace.contrib.internal.flask.patch import flask_version
 from ddtrace.contrib.internal.flask.patch import unpatch
-from ddtrace.internal.compat import StringIO
 from ddtrace.trace import Pin
 
 from . import BaseFlaskTestCase
