+++ tests/tracer/test_http.py
@@ -1,9 +1,9 @@
 import os
 import re
+from urllib import parse
 
 import pytest
 
-from ddtrace.internal.compat import parse
 from ddtrace.internal.utils.http import normalize_header_name
 from ddtrace.internal.utils.http import redact_url
 from ddtrace.internal.utils.http import strip_query_string
