+++ tests/tracer/test_rand.py
@@ -9,12 +9,12 @@ try:
 except ImportError:
     from multiprocessing.queues import SimpleQueue as MPQueue
 
+from queue import Queue
 import threading
 import time
 
 from ddtrace.internal import _rand
 from ddtrace.internal import forksafe
-from ddtrace.internal.compat import Queue
 from ddtrace.trace import tracer
